Added clock function for esp32
The esp32 connects via wifi to the NTP server and gets the up to date time. As the next step esp32 shows time on the integrated OLED. This function uses the library (NTPtimeESP32.zip) downloaded from the web. Everything works fine, no bugs tracked, i consider this part as done.